requirements.txt
numpy==1.24.3
Flask==2.3.2
tensorflow==2.13.0
Werkzeug==2.3.6
opencv-python==4.8.0.74
matplotlib==3.7.2
build==0.10.0
gunicorn==20.1.0
pylint==2.17.4
pyenchant==3.2.2
requests==2.31.0
pydantic==1.10.0
pdf2image
python-dotenv==1.0.0
roboflow==1.1.1
Pillow==9.0.1